refactor: move tolerant zip lookup to an Archive extension
The lookup tolerates `./` prefixes and percent-encoded paths, which is a
property of zip archives rather than of epub parsing, so it does not belong
on EpubParser.
