+++ iOS/UI/Reader/Page/ReaderPageView.swift
@@ -357,7 +357,7 @@ class ReaderPageView: UIView {
             do {
                 var data = Data()
                 let archive = try Archive(url: zipURL, accessMode: .read)
-                guard let entry = EpubParser.entry(in: archive, path: filePath) else {
+                guard let entry = archive.entry(at: filePath) else {
                     return nil
                 }
                 _ = try archive.extract(
