+++ iOS/UI/Reader/Readers/Text/Paged/ReaderPagedTextViewController.swift
@@ -339,7 +339,7 @@ class ReaderPagedTextViewController: BaseObservingViewController {
         do {
             var data = Data()
             let archive = try Archive(url: zipURL, accessMode: .read)
-            guard let entry = EpubParser.entry(in: archive, path: filePath) else {
+            guard let entry = archive.entry(at: filePath) else {
                 return nil
             }
             _ = try archive.extract(
